Fix illegal imports of `PasswordField`
I have moved `PasswordField` into the main chunk as it is almost
exclusively used there. The exception is in `Dialog.ts` where it is used
for request password dialog which I seperated into a separate file in
the main chunk.
