+++ src/misc/passwords/PasswordField.ts
@@ -1,13 +1,13 @@
 import m, { Children, Component, Vnode } from "mithril"
-import { ToggleButton } from "./buttons/ToggleButton.js"
-import { Icons } from "./icons/Icons.js"
-import { ButtonSize } from "./ButtonSize.js"
-import { Autocomplete, TextField, TextFieldAttrs, TextFieldType } from "./TextField.js"
-import { CompletenessIndicator } from "../CompletenessIndicator.js"
-import { isSecurePassword, scaleToVisualPasswordStrength } from "../../misc/passwords/PasswordUtils.js"
-import { Status, StatusField } from "./StatusField.js"
+import { ToggleButton } from "../../gui/base/buttons/ToggleButton.js"
+import { Icons } from "../../gui/base/icons/Icons.js"
+import { ButtonSize } from "../../gui/base/ButtonSize.js"
+import { Autocomplete, TextField, TextFieldAttrs, TextFieldType } from "../../gui/base/TextField.js"
+import { CompletenessIndicator } from "../../gui/CompletenessIndicator.js"
+import { isSecurePassword, scaleToVisualPasswordStrength } from "./PasswordUtils.js"
+import { Status, StatusField } from "../../gui/base/StatusField.js"
 import type { lazy } from "@tutao/tutanota-utils"
-import type { TranslationKey } from "../../misc/LanguageViewModel.js"
+import type { TranslationKey } from "../LanguageViewModel.js"
 
 type StatusSetting = Status | "auto"
